requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
defusedxml==0.7.1
Django==3.2
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-filter==22.1
djangorestframework==3.13.1
drf-nested-routers==0.93.4
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
urllib3==1.26.11
whitenoise==6.2.0
